[IMPROVE] Updated tooltips and icons from OmnichannelSection
