+++ apps/meteor/client/sidebar/sections/actions/OmnichannelLivechatToggle.tsx
@@ -1,0 +1,29 @@
+import { Sidebar } from '@rocket.chat/fuselage';
+import { useMutableCallback } from '@rocket.chat/fuselage-hooks';
+import { useMethod, useToastMessageDispatch, useTranslation } from '@rocket.chat/ui-contexts';
+import React, { ReactElement } from 'react';
+
+import { useOmnichannelAgentAvailable } from '../../../hooks/omnichannel/useOmnichannelAgentAvailable';
+
+export const OmnichannelLivechatToggle = (): ReactElement => {
+	const t = useTranslation();
+	const agentAvailable = useOmnichannelAgentAvailable();
+	const changeAgentStatus = useMethod('livechat:changeLivechatStatus');
+	const dispatchToastMessage = useToastMessageDispatch();
+
+	const props = {
+		title: agentAvailable ? t('Turn_off_answer_chats') : t('Turn_on_answer_chats'),
+		color: agentAvailable ? 'success' : undefined,
+		icon: agentAvailable ? 'message' : 'message-disabled',
+	} as const;
+
+	const handleAvailableStatusChange = useMutableCallback(async () => {
+		try {
+			await changeAgentStatus();
+		} catch (error: unknown) {
+			dispatchToastMessage({ type: 'error', message: error });
+		}
+	});
+
+	return <Sidebar.TopBar.Action {...props} onClick={handleAvailableStatusChange} />;
+};
