+++ apps/meteor/client/sidebar/sections/actions/index.ts
@@ -1,0 +1,6 @@
+export * from './OmnichannelCallDialPad';
+export * from './OmnichannelCallToggle';
+export * from './OmnichannelCallToggleError';
+export * from './OmnichannelCallToggleError';
+export * from './OmnichannelCallToggleReady';
+export * from './OmnichannelLivechatToggle';
